requirements.txt
flask
python-dotenv
requests
openai-whisper
huggingface_hub